+++ src/pages/News.js
@@ -34,9 +34,9 @@ export default function News() {
                     return (
                     <>
                     <div className="md:grid hidden rounded max shadow-md grid-cols-1 grid-rows-[1fr,40px] w-3/4 h-60 bg-white">
-                         <img className=" w-5/6 justify-self-center self-center px-1 border-2 row-start-1 row-end-2 col-start-1 col-end-2 h-5/6" src={element?.image?.key && `${API_BASE_URL}/api/v1/files/${element?.image?.key}`} ></img>
+                         <img className=" w-fit h-fit justify-self-center self-center px-1 border-2 row-start-1 row-end-2 col-start-1 col-end-2" src={element?.image?.key && `${API_BASE_URL}/api/v1/files/${element?.image?.key}`} ></img>
                          <h3 className=" px-1 place-self-center text-xl row-start-2 row-end-3 col-start-1 col-end-2">{element.name}</h3>
-                         <div className=" flex row-start-1 row-end-3 col-start-1 col-end-2 bg-slate-400 opacity-0 hover:opacity-80 justify-center items-center">
+                         <div className=" flex row-start-1 row-end-3 col-start-1 col-end-2 bg-slate-400 opacity-0 hover:opacity-100 justify-center items-center">
                             <button className=" hover:bg-sky-600 rounded p-2 opacity-100 tx text-white bg-sky-500 h-fit">
                                  <Link to={"/novedades/" + element.id}>MÁS INFORMACION</Link>
                              </button>
